pyproject.toml
[tool.poetry]
name = "intentional-walk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = ">=4.2.0,<5.0"
psycopg2 = "^2.9.1"
dj-database-url = "^0.5.0"
python-dotenv = "^0.11.0"
pytest-django = "^3.10.0"
coveralls = "^1.11.1"
coverage = "^5.0.3"
gunicorn = "^20.0.4"
python-dateutil = "^2.8.1"
whitenoise = "^5.0.1"
uuid = "^1.30"
faker = "^8.11.0"
black = "^22.8.0"
flake8 = "^5.0.4"
autoflake = "^1.5.3"
djangorestframework = "^3.14.0"
django-ninja = "^1.1.0"

[tool.poetry.dev-dependencies]
freezegun = "^1.1.0"
pytest = "^7.1.3"
libfaketime = "^2.0.0"
pytest-libfaketime = "^0.1.2"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
